requirements.txt
torchvision==0.10.1
matplotlib==3.4.3
numpy==1.21.2
torch==1.9.1
Pillow==8.3.2
